# backend/openapi/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Salud Nexus · API del centro sanitario",
    "summary": "Contrato HTTP del portal seguro de un único centro sanitario.",
    "description": "API REST versionada para el portal de pacientes de una instalación de Salud Nexus. El despliegue representa un solo centro; los identificadores internos de organización y centro se conservan únicamente para aislamiento, integridad referencial y auditoría. Los datos clínicos requieren una sesión de navegador Laravel Sanctum y todas las mutaciones están protegidas mediante CSRF. Los errores siguen RFC 9457 y nunca incluyen detalles clínicos o internos.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de plataforma Salud Nexus",
      "email": "platform@salud-nexus.example"
    },
    "license": {
      "name": "Propietario — uso interno autorizado",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 en el mismo origen que los portales Angular"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/patient/medications": {
      "get": {
        "tags": [
          "Medicación"
        ],
        "summary": "Listar la medicación del paciente",
        "description": "Devuelve registros profesionales y declaraciones del propio paciente, con elegibilidad calculada para renovación.",
        "operationId": "listPatientMedications",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de medicación.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationListEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          },
          "4XX": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/patient/medications/declarations": {
      "post": {
        "tags": [
          "Medicación"
        ],
        "summary": "Declarar medicación tomada por el paciente",
        "description": "Crea una declaración diferenciada de un registro profesional. No habilita renovación clínica y rechaza campos desconocidos.",
        "operationId": "declarePatientMedication",
        "security": [
          {
            "sessionCookie": [],
            "xsrfHeader": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclareMedicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Declaración creada o respuesta idempotente reproducida.",
            "headers": {
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayed"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          },
          "4XX": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Cookie HttpOnly, Secure y SameSite emitida por Laravel Sanctum para una sesión de navegador de primera parte."
      },
      "xsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "Valor URL-decodificado de la cookie XSRF-TOKEN. Es obligatorio para mutaciones y no sustituye a la sesión."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Clave opaca de 16 a 128 caracteres ASCII visibles. Debe reutilizarse solo para reintentar exactamente la misma operación.",
        "schema": {
          "type": "string",
          "pattern": "^[!-~]{16,128}$",
          "minLength": 16,
          "maxLength": 128
        }
      },
...
    },
    "headers": {
      "XRequestId": {
        "description": "Identificador de correlación validado o generado por el servidor.",
        "schema": {
          "$ref": "#/components/schemas/RequestId"
        }
      },
...
      "IdempotencyReplayed": {
        "description": "Indica si la respuesta procede de una solicitud idempotente anterior.",
        "schema": {
          "type": "string",
          "enum": [
            "true",
            "false"
          ]
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Error RFC 9457 neutral y correlacionable. El estado concreto depende del rechazo (400, 401, 403, 404, 405, 409, 413, 415, 419, 422, 429, 500 o 503).",
        "headers": {
          "Retry-After": {
            "description": "Segundos hasta un reintento cuando el servidor aplica limitación de tasa.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      },
      "UuidV7": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      },
      "RequestId": {
        "$ref": "#/components/schemas/Uuid"
      },
...
      "DateTime": {
        "type": "string",
        "format": "date-time"
      },
...
      "RequestMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "request_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://salud-nexus\\.example/problems/[a-z0-9-]+$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "pattern": "^/api/v1(?:/.*)?$"
          },
          "request_id": {
            "$ref": "#/components/schemas/RequestId"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "minItems": 1
            }
          }
        }
      },
...
      "Medication": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "source",
          "name",
          "presentation",
          "schedule_label",
          "status",
          "can_request_renewal",
          "renewal_request_status",
          "updated_at"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UuidV7"
          },
          "source": {
            "type": "string",
            "enum": [
              "professional_record",
              "patient_declaration"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "presentation": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          },
          "schedule_label": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "can_request_renewal": {
            "type": "boolean"
          },
          "renewal_request_status": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "submitted",
              null
            ]
          },
          "updated_at": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
      "MedicationEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Medication"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        }
      },
      "MedicationListEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medication"
            },
            "maxItems": 500
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        }
      },
      "DeclareMedicationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "schedule_label"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160,
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
          },
          "presentation": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120,
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
          },
          "schedule_label": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160,
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
          }
        }
      },
...
    }
  }
}